types and bugs
