+++ wamplify/src/app/types/types.ts
@@ -1,0 +1,16 @@
+export interface Assessment {
+    title: string,
+    weight: number,
+    score: number,
+  }
+  
+export interface Subject {
+    name: string,
+    code: string,
+    assessments: Assessment[],
+  }
+
+export interface SearchResult {
+    code: string,
+    name: string
+  }
